refactor(components/prometheus.py): rename to bkmonitor.py
